Function no_bs_fcs::data::create_dataframe

pub fn create_dataframe(
    column_titles: &[String],
    data: &[Vec<f64>]
) -> Result<DataFrame, PolarsError>

Creates a DataFrame from column titles and corresponding data vectors.

This function takes a vector of column titles and a vector of data vectors, converts each data vector to a Series with the corresponding column title, and combines them into a DataFrame.

Arguments

column_titles - A reference to a vector of strings representing the column titles.

data - A reference to a vector of vectors containing the data for each column.

Returns

A Result containing a DataFrame or a PolarsError if there is a mismatch in the number of columns and titles.

Errors

This function will return a PolarsError if:

The number of columns does not match the number of column titles.

Examples

use no_bs_fcs::data::create_dataframe;
use polars::prelude::*;
 
let column_titles = vec!["APC-A".to_string(), "FSC-W".to_string()];
let data = vec![vec![1.0, 2.0, 3.0], vec![4.0, 5.0, 6.0]];
let df = create_dataframe(&column_titles, &data).unwrap();
println!("{:?}", df);