Function no_bs_fcs::data::read_events

pub fn read_events<B: ByteOrder>(
    reader: &mut BufReader<&File>,
    data_type: &str,
    n_events: usize,
    param_idx: usize,
    metadata: &HashMap<String, String>
) -> Result<Vec<f64>, FcsError>

Reads the events data segment of the FCS file and returns a vector of f64 values.

This function reads the events data based on the specified data type (F, D, or I) and converts the raw data into a vector of f64 values. It handles different byte orders and bit depths as specified in the metadata.

Arguments

reader - A mutable reference to a BufReader wrapping the FCS file.

data_type - A string slice indicating the data type (‘F’ for float, ‘D’ for double, ‘I’ for integer).

n_events - The number of events to read.

param_idx - The parameter index for the current data segment.

metadata - A reference to a HashMap containing metadata from the FCS file.

Returns

A Result containing a vector of f64 values representing the events data, or an FcsError.

Errors

This function will return an FcsError if:

There is an I/O error during reading.

The specified data type is not supported.

The bits per parameter for integer data types is not supported or cannot be parsed.

Examples

use std::fs::File;
use std::io::BufReader;
use std::collections::HashMap;
use byteorder::LittleEndian;
use no_bs_fcs::text::read_metadata;
use no_bs_fcs::data::read_events;
 
let file = File::open("./examples/20200624 LEGENDplex_20200808 CMVMRC5 NY3 pDC.813537.fcs").unwrap();
let mut reader = BufReader::new(&file);
let metadata: HashMap<String, String> = read_metadata(&mut reader).unwrap();
let events = read_events::<LittleEndian>(&mut reader, "F", 1000, 1, &metadata).unwrap();
println!("{:?}", events);